--- src/foundry/client/core/index.d.mts
@@ -1,1 +1,0 @@
-import "./packages.d.mts";
